refactor: move Button to components/common
