+++ packages/app/src/components/common/index.tsx
@@ -1,0 +1,3 @@
+import Button from './Button/Button';
+
+export { Button };
